refactor: 테마 삭제 시 예약이 존재하면 예외 처리

+++ src/main/java/roomescape/service/ThemeService.java
@@ -2,6 +2,7 @@ package roomescape.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
+import roomescape.dao.ReservationDao;
 import roomescape.dao.ThemeDao;
 import roomescape.dto.ThemeRequest;
 import roomescape.dto.ThemeResponse;
@@ -9,9 +10,11 @@ import roomescape.model.Theme;
 
 @Service
 public class ThemeService {
+    private final ReservationDao reservationDao;
     private final ThemeDao themeDao;
 
-    public ThemeService(ThemeDao themeDao) {
+    public ThemeService(ReservationDao reservationDao, ThemeDao themeDao) {
+        this.reservationDao = reservationDao;
         this.themeDao = themeDao;
     }
 
@@ -27,6 +30,9 @@ public class ThemeService {
     }
 
     public void deleteTheme(Long id) {
+        if (reservationDao.isExistByThemeId(id)) {
+            throw new IllegalStateException("예약이 존재하여 삭제할 수 없습니다.");
+        }
         boolean isDeleted = themeDao.deleteById(id);
         if (!isDeleted) {
             throw new IllegalArgumentException("해당하는 ID가 없습니다.");
